cleaning up a bit

+++ src/main/java/net/dialingspoon/multicount/interfaces/AdvancementAdditions.java
@@ -1,5 +1,5 @@
 package net.dialingspoon.multicount.interfaces;
 
 public interface AdvancementAdditions {
-    void saveSecond(byte next, byte old);
+    void saveSecond(int next, int old);
 }
