+++ src/main/java/net/dialingspoon/multicount/interfaces/StatHandlerAdditions.java
@@ -1,5 +1,5 @@
 package net.dialingspoon.multicount.interfaces;
 
 public interface StatHandlerAdditions {
-    void saveSecond(byte next, byte old);
+    void saveSecond(int next, int old);
 }
